Homepage carousels update
IA and IoT projects have been moved to carousels.

A new drop shadow effect is present on the project cards.

README added.

I've wrote up additional context for all the projects. I'm looking to rework the read more on the cards, but that'll be later.

Next update tackles whatever this did to the mobile view.